Move the JPanelListRenderer into SIF.
