+++ bundles/wps/wps-client/src/main/java/org/orbisgis/wpsclient/view/ui/dataui/DataStoreUI.java
@@ -41,9 +41,9 @@ import net.miginfocom.swing.MigLayout;
 import net.opengis.wps._2_0.DescriptionType;
 import net.opengis.wps._2_0.InputDescriptionType;
 import org.orbisgis.sif.common.ContainerItem;
+import org.orbisgis.sif.components.renderers.JPanelListRenderer;
 import org.orbisgis.wpsclient.WpsClientImpl;
 import org.orbisgis.wpsclient.view.utils.ToolBoxIcon;
-import org.orbisgis.wpsclient.view.utils.sif.JPanelListRenderer;
 import org.orbisgis.wpsservice.model.*;
 import org.xnap.commons.i18n.I18n;
 import org.xnap.commons.i18n.I18nFactory;
